move common handlers
