+++ handlers/common/static_handlers.go
@@ -1,4 +1,4 @@
-package handlers
+package common
 
 import (
 	"bytes"
