+++ handlers/images/routes.go
@@ -16,10 +16,10 @@ import (
 	"github.com/gorilla/mux"
 
 	corecommon "github.com/arran4/goa4web/core/common"
+	handlerspkg "github.com/arran4/goa4web/handlers/common"
 	hcommon "github.com/arran4/goa4web/handlers/common"
 	db "github.com/arran4/goa4web/internal/db"
 	router "github.com/arran4/goa4web/internal/router"
-	handlerspkg "github.com/arran4/goa4web/pkg/handlers"
 	imagesign "github.com/arran4/goa4web/pkg/images"
 	"github.com/arran4/goa4web/pkg/upload"
 	"github.com/arran4/goa4web/runtimeconfig"
