+++ handlers/user/routes.go
@@ -7,7 +7,6 @@ import (
 	auth "github.com/arran4/goa4web/handlers/auth"
 	"github.com/arran4/goa4web/handlers/common"
 	router "github.com/arran4/goa4web/internal/router"
-	"github.com/arran4/goa4web/pkg/handlers"
 )
 
 // RegisterRoutes attaches user account endpoints to the router.
@@ -43,8 +42,8 @@ func RegisterRoutes(r *mux.Router) {
 	ur.HandleFunc("/subscriptions/delete", userSubscriptionsDeleteAction).Methods(http.MethodPost).MatcherFunc(auth.RequiresAnAccount()).MatcherFunc(common.TaskMatcher(common.TaskDelete))
 
 	// legacy redirects
-	r.HandleFunc("/user/lang", handlers.RedirectPermanent("/usr/lang"))
-	r.HandleFunc("/user/email", handlers.RedirectPermanent("/usr/email"))
+	r.HandleFunc("/user/lang", common.RedirectPermanent("/usr/lang"))
+	r.HandleFunc("/user/email", common.RedirectPermanent("/usr/email"))
 }
 
 // Register registers the user router module.
